Wrap visitors component
